22 Jun 2004  Billy Biggs <vektor@dumbterm.net>
  * tvtime/src/cpuinfo.c: Fix for compiling the tvtime executable with
      -fPIC.  Patch from Than Ngo <than@redhat.com>.
  * tvtime/src/wine/win32.c: Apparently this fixes a compile error on
      gcc-3.4.  Patch from Than Ngo <than@redhat.com>.
  * tvtime/AUTHORS: Mention Than Ngo in the AUTHORS file.

+++ src/wine/win32.c
@@ -2611,7 +2611,7 @@ static int WINAPI expGetWindowThreadProcessId(HWND win, int *pid_data)
     dbgprintf("GetWindowThreadProcessId(0x%x, 0x%x) => %d\n",
 	win, pid_data, tid);
     if (pid_data)
-	(int)*pid_data = tid;
+	*pid_data = tid;
     return tid;
 }
